store: Move connection_pool.rs to its own module
No functional change
