+++ store/postgres/src/chain_head_listener.rs
@@ -11,8 +11,8 @@ use std::sync::Arc;
 use lazy_static::lazy_static;
 
 use crate::{
-    connection_pool::ConnectionPool,
     notification_listener::{JsonNotification, NotificationListener, SafeChannelName},
+    pool::ConnectionPool,
     NotificationSender,
 };
 use graph::blockchain::ChainHeadUpdateListener as ChainHeadUpdateListenerTrait;
