+++ store/postgres/src/lib.rs
@@ -17,7 +17,6 @@ mod block_store;
 mod catalog;
 mod chain_head_listener;
 mod chain_store;
-mod connection_pool;
 mod copy;
 mod deployment;
 mod deployment_store;
@@ -27,6 +26,7 @@ mod fork;
 mod functions;
 mod jobs;
 mod notification_listener;
+mod pool;
 mod primary;
 pub mod query_store;
 mod relational;
@@ -60,10 +60,10 @@ pub use self::block_store::BlockStore;
 pub use self::block_store::ChainStatus;
 pub use self::chain_head_listener::ChainHeadUpdateListener;
 pub use self::chain_store::{ChainStore, ChainStoreMetrics, Storage};
-pub use self::connection_pool::{ConnectionPool, ForeignServer, PoolCoordinator, PoolRole};
 pub use self::detail::DeploymentDetail;
 pub use self::jobs::register as register_jobs;
 pub use self::notification_listener::NotificationSender;
+pub use self::pool::{ConnectionPool, ForeignServer, PoolCoordinator, PoolRole};
 pub use self::primary::{db_version, UnusedDeployment};
 pub use self::store::Store;
 pub use self::store_events::SubscriptionManager;
